use double precision during aggregation.

+++ aero_sat/aero_load_calculator/gpu_aero_load_calculator.h
@@ -11,8 +11,8 @@
 #include <memory>
 
 struct ForceTorqueData {
-    glm::vec4 force;
-    glm::vec4 torque;
+    glm::dvec4 force;
+    glm::dvec4 torque;
 };
 
 class GPUAeroLoadCalculator: public IAeroLoadCalculator {
